Backend/requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
google-generativeai==0.8.3
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
tzdata==2024.2
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0  
python-dotenv==1.0.0 
